Merge existing options with mocking WithOptions
The behaviour with Ops in normal usage is to merge any options specified
in WithOptions with the Op's existing Options. This was not being
followed in mocks. This commit fixes this by making sure we always merge
with any existing options, thus preventing the mocked WithOptions
unsetting any options already set.

This commit also adds a test case to prevent regressions